% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_parse_metadata.R
\name{addquotes}
\alias{addquotes}
\title{Helper for metadata parser, adds quotes around values in a string}
\usage{
addquotes(txt)
}
\arguments{
\item{txt}{Text string to add quotes around}
}
\value{
Text string with added quotes
}
\description{
Helper for metadata parser, adds quotes around values in a string
}
\examples{
addquotes('2021')
addquotes(c("2021","2022","2023"))

}
\seealso{
\code{\link[=splitlist]{splitlist()}} for adding quotes to each element in a vector and joining to single string, and
\code{\link[=px_parse_metadata]{px_parse_metadata()}} for parsing metadata to text
}
