% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_utils.R
\name{convert_data_to_final}
\alias{convert_data_to_final}
\title{Convert data in long format into same format according to STUB and HEADING.
The dataframe column names are assumed to come from the main language defined in the LANGUAGE keyword.}
\usage{
convert_data_to_final(.metadata_df, .data)
}
\arguments{
\item{.metadata_df}{Metadata tibble}

\item{.data}{Data tibble in long format}
}
\value{
Returns tibble in wide format according to order of STUB and HEADING
}
\description{
Convert data in long format into same format according to STUB and HEADING.
The dataframe column names are assumed to come from the main language defined in the LANGUAGE keyword.
}
