% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_read.R
\name{create_px_object_from_dataframe}
\alias{create_px_object_from_dataframe}
\title{Create px object from data frame}
\usage{
create_px_object_from_dataframe(.data)
}
\arguments{
\item{.data}{data frame or tibble}
}
\value{
returns a px object of class pxR2, ie a list of dataframes containing metadata and data
}
\description{
Create px object from data frame
}
\examples{
pxR2::create_px_object_from_dataframe(ex_data)

}
\keyword{internal}
