% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_utils.R
\name{get_value_by_keyword}
\alias{get_value_by_keyword}
\title{Get value from metadata tibble by keyword}
\usage{
get_value_by_keyword(.metadata_df, .key)
}
\arguments{
\item{.metadata_df}{metadata tibble to extract from}

\item{.key}{keyword you want to get value from}
}
\value{
returns value(s) matching the keyword
}
\description{
Get value from metadata tibble by keyword
}
\examples{
get_value_by_keyword(meta_example, "STUB")
}
