% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_utils.R
\name{keyword_exists}
\alias{keyword_exists}
\title{Check if keyword exists in metadata tibble}
\usage{
keyword_exists(.metadata_df, .key)
}
\arguments{
\item{.metadata_df}{metadata tibble}

\item{.key}{keyword to check if it exists}
}
\value{
returns TRUE if keyword exists
}
\description{
Check if keyword exists in metadata tibble
}
\examples{
keyword_exists(meta_example, "DECIMALS")
keyword_exists(meta_example, "SHOWDECIMALS")

}
