% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_example}
\alias{meta_example}
\title{Example metadata tibble}
\format{
\subsection{\code{meta_example}}{

A data frame with 18 rows and 5 columns:
\describe{
\item{keyword}{PX keyword in uppercase.}
\item{language}{Language. If only one language exists it can be omitted,
otherwise language should be defined where a translation is needed.}
\item{varname}{The variable name. Used if a keyword refers to a specific variable.
For example, NOTE("age") is a footnote for the varname "age".}
\item{valname}{The value name. Used if a keyword refers to a specific variable value,
for example VALUENOTE("age","0-15") is a footnote for the age group 0-15 years.}
\item{value}{The value to show for each keyword.}
}
}
}
\usage{
meta_example
}
\description{
An example metadata tibble needed for creating a px-object.
The metadata tibble accepts all PX keywords according to the ´specs´ except for
CELLNOTE(X) and DATANOTECELL.

For more information on available px keyword, check out the dataset \code{specs} or
\code{show_keyword_help(keyword)}.
}
\keyword{datasets}
