% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_create.R
\name{px_create}
\alias{px_create}
\title{Creates new px object from a tibble/data frame.
Creates all necessary metadata and generates VALUE keywords dynamically based on the variable levels in the data frame.}
\usage{
px_create(
  .data,
  meta_csv_path = NULL,
  codes_csv_path = NULL,
  stub,
  heading,
  time_variable = NULL,
  time_scale = NULL,
  matrix,
  subject_area,
  subject_code,
  units,
  contents,
  decimals = 1,
  showdecimals = 1,
  language = "en",
  charset = "ANSI",
  axis_version = "2013",
  codepage = "iso-8859-15",
  creation_date = format(Sys.time(), "\%Y\%m\%d \%H:\%M"),
  last_updated = format(Sys.time(), "\%Y\%m\%d \%H:\%M"),
  contact = NULL,
  source = NULL,
  note = NULL
)
}
\arguments{
\item{.data}{Mandatory. Dataframe/tibble in long format with one column called 'value'.}

\item{meta_csv_path}{Path to metadata csv (;-delimited)}

\item{codes_csv_path}{Path to codes csv (;-delimited)}

\item{stub}{Mandatory. Variables to display along the rows. Supply as a comma separated string, for example 'country,age,'}

\item{heading}{Mandatory. Variables to display along the columns.}

\item{time_variable}{Name of the time variable, if it exists.}

\item{time_scale}{If time_variable exist, you must set the time scale. Choose from 'annual', 'halfyear', 'quarterly', 'monthly' or 'weekly'.
Choose annual format if the time variable is formatted as CCYY (C for century, Y for year).
Choose halfyear format if the time variable is formatted as CCYYH (H is 1 or 2).
Choose quarterly format if the time variable is formatted as CCYYQ (Q is 1-4).
Choose monthly format if the time variable is formatted as CCYYMM (M is 1-12).
Choose weekly format if the time variable is formatted as CCYYWW (M is 1-52).}

\item{matrix}{Mandatory. The name of the matrix, should be the same as the file name, typically in uppercase. Max length 20 characters.}

\item{subject_area}{Mandatory. The name of the subject are the matrix should be categorized under.}

\item{subject_code}{Mandatory. Code for the subject are. Typically upper case acronym, max length of 20 characters.}

\item{units}{Mandatory. The specific type of units used in the table in plain text, for example: 'Number of applications', 'Ton', 'Index', etc.}

\item{contents}{Mandatory. The first part of a title excluding the variables used.
For example, if the title is 'Number of cats in Europe by country, age and time', the CONTENT is 'Number of cats in Europe'.
The last part of the title, 'by country, age, and time', will be automatically generated and written to TITLE.}

\item{decimals}{Mandatory. Integer between 0-15 (0-6 if SHOWDECIMALSis not included). Indicates how many decimals will be saved in the px-file.}

\item{showdecimals}{The number of decimals to be shown in the table, integer between 0-6.
Must be the same or smaller than the number stored as indicated by the keyword DECIMALS.
If SHOWDECIMALS is not stated in the file the number stated by DECIMALS will be used.}

\item{language}{The language for the px-file, 'sv' for Swedish, 'en' for English, etc.
If the keyword is used, the words for “and” and “by” are read from the text file of that language which is used to generate the TITLE.}

\item{charset}{CHARSET=”ANSI”; indicates that the texts in the file are written in Windows format.
If the keyword is missing it means that the texts in the file are in DOS format.}

\item{axis_version}{Version number for PC-Axis, deafault is the latest version 2013. Is read and saved but otherwise not used.}

\item{codepage}{Is used when creating XML format to get correct characters. Default iso8859-1. Max 20 chars}

\item{creation_date}{Date when file was created. Written in format CCYYMMDD hh:mm, e.g. ”19960612 14:20”. Is shown together with footnotes.}

\item{last_updated}{Date and time for latest update format CCYYMMDD hh:mm. Example ”19960528 11:35”.}

\item{contact}{States contact information for the matrix such as a email and is shown in the footnotes.
Can for example be written in the form of: name, organization, telephone, fax, e-mail.
Several persons can be stated in the same text string and are then divided by the #-sign.}

\item{source}{States the organization which is responsible for the statistics or the sources used. Is shown with the footnote.
If multiple sources are used, use # between each source. For example 'Statistics Sweden#Statistics Finland'.}

\item{note}{General footnote for the table.}
}
\value{
Returns a list with two entries, one containing the metadata information and one for the data.
}
\description{
Creates new px object from a tibble/data frame.
Creates all necessary metadata and generates VALUE keywords dynamically based on the variable levels in the data frame.
}
\examples{
px_obj <- px_create(.data = ex_data,
          stub = "sex,age",
          heading = "time",
          time_variable = "time",
          time_scale="annual",
          matrix = "TEST01",
          subject_area = "Test",
          subject_code = "T",
          units = "Antal timmar",
          contents = "Genomsnitt antal tittartimmar av The Simpsons",
          decimals = 1,
          language = "en"
)

px_obj

}
