% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_utils.R
\name{px_generate_dynamic_title}
\alias{px_generate_dynamic_title}
\title{Generate dynamic TITLE based on LANGUAGE, STUB and HEADING}
\usage{
px_generate_dynamic_title(.metadata_df)
}
\arguments{
\item{.metadata_df}{metadata tibble to extract from}
}
\value{
returns single string as dynamically generated title.
}
\description{
Generate dynamic TITLE based on LANGUAGE, STUB and HEADING
}
\examples{
\dontrun{
px_generate_dynamic_title(meta_example)
px_generate_dynamic_title(metadata_example_multilingual)
}
}
\keyword{internal}
