% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_utils.R
\name{px_get_languages}
\alias{px_get_languages}
\title{Helper function to extract LANGUAGES from metadata tibble as a character vector.
If not defined, NA_character_ is returned.}
\usage{
px_get_languages(.metadata_df)
}
\arguments{
\item{.metadata_df}{Metadata tibble}
}
\value{
Returns character vector of languages or NA_character_ if undefined.
}
\description{
Helper function to extract LANGUAGES from metadata tibble as a character vector.
If not defined, NA_character_ is returned.
}
\examples{
\dontrun{
px_get_languages(meta_example)
px_get_languages(metadata_example_multilingual)
}
}
\keyword{internal}
