% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_utils.R
\name{px_get_main_language}
\alias{px_get_main_language}
\title{Helper function to extract main language from metadata tibble, if it exists.
If LANGUAGE keyword exists, that is assumed to be the main language.
If LANGUAGE keyword doesn't exist, but LANGUAGES does, the first language of LANGUAGES is assumed to be the main language.}
\usage{
px_get_main_language(.metadata_df)
}
\arguments{
\item{.metadata_df}{Metadata tibble.}
}
\value{
Returns the main language as character vector.
}
\description{
Helper function to extract main language from metadata tibble, if it exists.
If LANGUAGE keyword exists, that is assumed to be the main language.
If LANGUAGE keyword doesn't exist, but LANGUAGES does, the first language of LANGUAGES is assumed to be the main language.
}
\examples{
px_get_main_language(meta_example)
px_get_main_language(metadata_example_multilingual)

px_meta_init_unempty(keyword="CHARSET", NA, NA, NA, value = "ANSI") |>
px_get_main_language()

px_meta_init_unempty(keyword="LANGUAGES", NA, NA, NA, value = "sv,en") |>
px_get_main_language()

px_meta_init_unempty(keyword="LANGUAGE", NA, NA, NA, value = "en") |>
px_get_main_language()

}
