% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_utils.R
\name{px_get_values_from_data}
\alias{px_get_values_from_data}
\title{Extract values from dataframe/tibble into metadata tibble.}
\usage{
px_get_values_from_data(.data, as_list = FALSE)
}
\arguments{
\item{.data}{The dataframe or tibble to extract unique variable values from.}

\item{as_list}{If TRUE, all unique variable values are returned as a list.}
}
\value{
Returns metadata tibble or list with extracted unique variable values
}
\description{
Extract values from dataframe/tibble into metadata tibble.
}
\examples{
\dontrun{
px_get_values_from_data(ex_data)
px_get_values_from_data(ex_data, as_list = TRUE)
}
}
\keyword{internal}
