% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_meta_handle.R
\name{px_meta_add_keyword}
\alias{px_meta_add_keyword}
\title{Add new row of keywords to metadata tibble}
\usage{
px_meta_add_keyword(
  metadata,
  keyword,
  language = NA,
  varname = NA,
  valname = NA,
  value
)
}
\arguments{
\item{metadata}{Metadata tibble to add to.}

\item{keyword}{Mandatory. PX keyword in capital letters.}

\item{language}{Language code, for example 'en', 'sv', etc.}

\item{varname}{Name of variable you want to comment.}

\item{valname}{Name of the value you want to comment.}

\item{value}{Mandatory. The text you want to add.}
}
\value{
Metadata tibble
}
\description{
Add new row of keywords to metadata tibble
}
\examples{
px_meta_add_keyword(meta_example, "VALUENOTE", NA, "typ", "typ value", "Preliminary figures")
}
