% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_meta_handle.R
\name{px_meta_add_timeval}
\alias{px_meta_add_timeval}
\title{Add time value information to px metadata tibble}
\usage{
px_meta_add_timeval(.metadata_df, time_variable, time_scale)
}
\arguments{
\item{.metadata_df}{Metadata tibble to add to.}

\item{time_variable}{Name of variable containing time info, for example 'time', 'år', etc.}

\item{time_scale}{Choose from 'annual', 'halfyear', 'quarterly', 'monthly', 'weekly'.
Choose annual format if the time variable is formatted as CCYY (C for century, Y for year).
Choose halfyear format if the time variable is formatted as CCYYH (H is 1 or 2).
Choose quarterly format if the time variable is formatted as CCYYQ (Q is 1-4).
Choose monthly format if the time variable is formatted as CCYYMM (M is 1-12).
Choose weekly format if the time variable is formatted as CCYYWW (M is 1-52).}
}
\value{
Metadata tibble
}
\description{
Add time value information to px metadata tibble
}
