% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_meta_handle.R
\name{px_meta_init_empty}
\alias{px_meta_init_empty}
\title{Initialize empty px metadata tibble}
\usage{
px_meta_init_empty()
}
\value{
Returns empty px metadata tibble
}
\description{
Initialize empty px metadata tibble
}
\examples{
px_meta_init_empty()
}
