% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_meta_handle.R
\name{px_meta_init_unempty}
\alias{px_meta_init_unempty}
\title{Initialize nonempty px metadata tibble with arguments}
\usage{
px_meta_init_unempty(keyword, language, varname, valname, value)
}
\arguments{
\item{keyword}{Mandatory. PX keyword in capital letters.}

\item{language}{Language code, for example 'en', 'sv', etc.}

\item{varname}{Name of variable you want to comment.}

\item{valname}{Name of the value you want to comment.}

\item{value}{Mandatory. The text you want to add.}
}
\value{
Filled px metadata tibble
}
\description{
Initialize nonempty px metadata tibble with arguments
}
\examples{
px_meta_init_unempty("VALUENOTE", "", "age", "15-20 years", "Preliminary figures")
px_meta_init_unempty("valuenote", "", "age", "15-20 years", "Preliminary figures")
# will throw error if keyword is not valid according to PX Specifications:
px_meta_init_unempty("HEJ", "", "age", "15-20 years", "Preliminary figures")
}
