% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{px_parse}
\alias{px_parse}
\title{Parse a px file}
\usage{
px_parse(filePath, only_meta = FALSE, debug = FALSE)
}
\arguments{
\item{filePath}{path to PX file ending in .px or .PX}

\item{only_meta}{Boolean, set to true if you only want to extract the metadata.}

\item{debug}{Boolean, set to true if you want to debug the parser step-by-step.}
}
\value{
Returns a list of metadata and data.
}
\description{
This function read a px file, parsing data and metadata and warnins if
the PX file is malformatted.
}
