% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_parse_metadata.R
\name{px_parse_metadata}
\alias{px_parse_metadata}
\title{Parse px metadata to px-formatted text.}
\usage{
px_parse_metadata(.metadata, full_dataframe = FALSE)
}
\arguments{
\item{.metadata}{Metadata as input to be converted to px-formatted text.}

\item{full_dataframe}{If TRUE, return entire dataframe. Useful for debugging. Default FALSE.}
}
\value{
Px-formatted text
}
\description{
Parse px metadata to px-formatted text.
}
\keyword{internal}
