% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_read.R
\name{px_read}
\alias{px_read}
\title{Create px object from dataframe or px file}
\usage{
px_read(input)
}
\arguments{
\item{input}{either a datarframe/tibble or a path to a px file ending in \code{.px} or \code{.PX}}
}
\value{
returns a px object of class pxR2
}
\description{
Create px object from dataframe or px file
}
