% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_write.R
\name{px_write}
\alias{px_write}
\title{Write px object to a PX file}
\usage{
px_write(.px_object, file)
}
\arguments{
\item{.px_object}{px object as a list of a metadata tibble with a data tibble.}

\item{file}{destination file, ending in .px}
}
\value{
saves to file
}
\description{
Write px object to a PX file
}
