% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxR2-package.R
\docType{package}
\name{pxR2-package}
\alias{pxR2}
\alias{pxR2-package}
\title{pxR2: Read and Write PX Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of pxR2 is to provide a faster and easier way to read and write px-data (PC-AXIS), inspired by the packages pxR and readr. It is designed to correctly parse data and metadata and easy integration with other px-related programs such as PXjob and PXEdit.
}
\author{
\strong{Maintainer}: Emil Westin \email{emilwestin@gmail.com}

}
\keyword{internal}
