% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_meta_utils.R
\name{show_keyword_help}
\alias{show_keyword_help}
\title{Show keyword help information}
\usage{
show_keyword_help(keyword)
}
\arguments{
\item{keyword}{PX keyword}
}
\value{
returns formatted string of help information
}
\description{
This function provides details on a keyword according to the PX-file format specification and reads from \code{specs}.
It displays:
}
\details{
\itemize{
\item is the keyword mandatory/language dependent?
\item can values of the keyword span over multiple lines? (Multiline)
\item type (text/integer/other)
\item max character length allowed for values
\item default value (if any)
\item a keyword description
}
}
\examples{
show_keyword_help("TIMEVAL")
show_keyword_help("NOTE")
show_keyword_help("CELLNOTE")
show_keyword_help("LANGUAGE")
show_keyword_help("CODES")
show_keyword_help("DATA")

}
\seealso{
\link{specs} for PX file format specification tibble
or \url{https://www.scb.se/globalassets/vara-tjanster/px-programmen/px-file_format_specification_2013.pdf} for original documentation by Statistics Sweden.
}
