% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_utils.R
\name{split_commas}
\alias{split_commas}
\title{Split commas in for example stub}
\usage{
split_commas(.values)
}
\arguments{
\item{.values}{single string with commas}
}
\value{
splitted vector
}
\description{
Split commas in for example stub
}
\examples{
get_value_by_keyword(meta_example, "STUB") |> split_commas()
get_value_by_keyword(metadata_example_multilingual, "STUB") |> map(split_commas)
}
