% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_parse_metadata.R
\name{px_parse_metadata}
\alias{px_parse_metadata}
\title{Parse px metadata tibble to px-formatted text.}
\usage{
px_parse_metadata(.metadata_df)
}
\arguments{
\item{.metadata_df}{Metadata tibble as input to be converted to px-formatted text.}
}
\value{
Px-formatted text
}
\description{
Parse px metadata tibble to px-formatted text.
}
\examples{

px_parse_metadata(meta_example)
px_parse_metadata(metadata_example_multilingual)

}
