% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_write.R
\name{px_write}
\alias{px_write}
\title{Write px object to a PX file}
\usage{
px_write(.px_object, file)
}
\arguments{
\item{.px_object}{px object as a list of a metadata tibble with a data tibble.}

\item{file}{destination file, ending in .px}
}
\value{
saves to file
}
\description{
Write px object to a PX file
}
\examples{
\dontrun{
px_obj <- px_create(tt,
stub = "sex,age",
heading = "time",
time_variable = "time",
time_scale = "annual",
matrix = "TEST01",
subject_area = "Forestry",
subject_code = "F",
units = "Number",
contents = "Number of trees",
decimals = 1,
language = "en"
)

px_write(px_obf, "./test.px")
}
}
